update the library membership lastEmailSent for digest emails
